# OpenAPI/Specs/FlightInformationAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Flight Information API",
    "summary": "Flight Information API Ver 1.0.0",
    "description": "The API provides endpoints to search, retrieve, create, update, and delete flight records. It supports querying flights by airline, route, and date range, and returns details such as flight number, departure/arrival airports, times, and current status."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/flights": {
      "get": {
        "tags": [
          "Flights"
        ],
        "summary": "GET all flights endpoint",
        "description": "Retrieve all flights.",
        "operationId": "GetAllFlights",
        "responses": {
          "200": {
            "description": "Retrieve all flights successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightDetail"
                  }
                },
                "examples": {
                  "Example 1": {
                    "value": [
                      {
                        "id": 1,
                        "flightNumber": "NZ123",
                        "airline": "Air New Zealand",
                        "departureAirport": "AKL",
                        "arrivalAirport": "WLG",
                        "departureTime": "2025-11-26T10:00:00Z",
                        "arrivalTime": "2025-11-26T11:00:00Z",
                        "status": "Scheduled"
                      },
                      {
                        "id": 2,
                        "flightNumber": "NZ456",
                        "airline": "Air New Zealand",
                        "departureAirport": "WLG",
                        "arrivalAirport": "AKL",
                        "departureTime": "2025-12-26T10:00:00Z",
                        "arrivalTime": "2025-12-26T11:00:00Z",
                        "status": "Scheduled"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "/api"
          }
        ]
      },
      "post": {
        "tags": [
          "Flights"
        ],
        "summary": "CREATE new flight endpoint",
        "description": "Creates a new flight with the provided details. On success, returns the created flight and the Location header pointing to the new resource.",
        "operationId": "CreateFlight",
        "requestBody": {
          "description": "Flight details used to create a new flight.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightDetail"
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "flightNumber": "NZ123",
                    "airline": "Air New Zealand",
                    "departureAirport": "AKL",
                    "arrivalAirport": "WLG",
                    "departureTime": "2025-11-26T10:00:00Z",
                    "arrivalTime": "2025-11-26T11:00:00Z",
                    "status": "Scheduled"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Flight created successfully.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "example": "/flights/123"
                },
                "description": "URL of the newly created flight resource."
              }
            }
          },
          "400": {
            "description": "The request is invalid or fails validation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "error": "ValidationFailed",
                      "message": "One or more validation errors occurred.",
                      "details": [
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "servers": [
          {
            "url": "/api"
          }
        ]
      }
    },
    "/flights/{id}": {
      "get": {
        "tags": [
          "Flights"
        ],
        "summary": "GET flight detail by id endpoint",
        "description": "Retrieve a specific flight detail by ID.",
        "operationId": "GetFlight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the flight to retrieve.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieve a specific flight by ID successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightDetail"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "id": 1,
                      "flightNumber": "NZ123",
                      "airline": "Air New Zealand",
                      "departureAirport": "AKL",
                      "arrivalAirport": "WLG",
                      "departureTime": "2025-11-26T10:00:00Z",
                      "arrivalTime": "2025-11-26T11:00:00Z",
                      "status": "Scheduled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Flight is not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "error": "NotFound",
                      "message": "Flight with the specified ID was not found.",
                      "details": [
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "/api"
          }
        ]
      },
      "put": {
        "tags": [
          "Flights"
        ],
        "summary": "UPDATE a flight detail endpoint",
        "description": "Update an existing flight detail.",
        "operationId": "UpdateFlight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the flight to retrieve.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "description": "Flight details used to update an existing flight.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlightDetail"
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "id": 1,
                    "flightNumber": "NZ123",
                    "airline": "Air New Zealand",
                    "departureAirport": "AKL",
                    "arrivalAirport": "WLG",
                    "departureTime": "2025-11-26T10:00:00Z",
                    "arrivalTime": "2025-11-26T11:00:00Z",
                    "status": "Scheduled"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightDetail"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "id": 1,
                      "flightNumber": "NZ123",
                      "airline": "Air New Zealand",
                      "departureAirport": "AKL",
                      "arrivalAirport": "WLG",
                      "departureTime": "2025-11-26T10:00:00Z",
                      "arrivalTime": "2025-11-26T11:00:00Z",
                      "status": "Scheduled"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The request is invalid or fails validation (for example, missing required fields or invalid date/time values).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "error": "ValidationFailed",
                      "message": "One or more validation errors occurred.",
                      "details": [
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Flight with the specified ID was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "error": "NotFound",
                      "message": "Flight with the specified ID was not found.",
                      "details": [
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The update could not be applied due to a conflict, such as a concurrency violation or a business rule conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "error": "Conflict",
                      "message": "Concurrentcy conflict, two or more users update the same flight at the same time.",
                      "details": [
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "/api"
          }
        ]
      },
      "delete": {
        "tags": [
          "Flights"
        ],
        "summary": "DELETE a flight endpoint",
        "description": "Delete a flight.",
        "operationId": "DeleteFlight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the flight to retrieve.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlightDetail"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "id": 1,
                      "flightNumber": "NZ123",
                      "airline": "Air New Zealand",
                      "departureAirport": "AKL",
                      "arrivalAirport": "WLG",
                      "departureTime": "2025-11-26T10:00:00Z",
                      "arrivalTime": "2025-11-26T11:00:00Z",
                      "status": "Scheduled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Flight is not Found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "error": "NotFound",
                      "message": "Flight with the specified ID was not found.",
                      "details": [
                        "string"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "/api"
          }
        ]
      }
    },
    "/flights/search": {
      "get": {
        "tags": [
          "Flights"
        ],
        "summary": "SEARCH flights by keys endpoint",
        "description": "Searches for flights that match the specified criteria. Use the searchKeys query parameter to filter by airline, departure and arrival airports, and an optional departure date range. Returns a list of matching flights when found, or 204 (No Content) if no flights match.",
        "operationId": "SearchFlights",
        "parameters": [
          {
            "name": "searchKeys",
            "in": "query",
            "description": "Search criteria for filtering flights.",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "$ref": "#/components/schemas/SearchKeys"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of flights that match the provided search criteria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FlightDetail"
                  }
                },
                "examples": {
                  "Example 1": {
                    "value": [
                      {
                        "id": 1,
                        "flightNumber": "NZ123",
                        "airline": "Air New Zealand",
                        "departureAirport": "AKL",
                        "arrivalAirport": "WLG",
                        "departureTime": "2025-11-26T10:00:00Z",
                        "arrivalTime": "2025-11-26T11:00:00Z",
                        "status": "Scheduled"
                      },
                      {
                        "id": 2,
                        "flightNumber": "NZ456",
                        "airline": "Air New Zealand",
                        "departureAirport": "WLG",
                        "arrivalAirport": "AKL",
                        "departureTime": "2025-12-26T10:00:00Z",
                        "arrivalTime": "2025-12-26T11:00:00Z",
                        "status": "Scheduled"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "servers": [
          {
            "url": "/api"
          }
        ]
      }
    }
  },
  "tags": [
    {
      "name": "Flights",
      "description": "Operations for searching and retrieving flight information"
    }
  ],
  "components": {
    "schemas": {
      "FlightDetail": {
        "type": "object",
        "title": "Flight Detail",
        "description": "A representation of a flight information.",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Flight id."
          },
          "flightNumber": {
            "type": "string",
            "description": "Flight number."
          },
          "airline": {
            "type": "string",
            "description": "Name of the airline. "
          },
          "departureAirport": {
            "type": "string",
            "description": "Departure airport."
          },
          "arrivalAirport": {
            "type": "string",
            "description": "Destination airport."
          },
          "departureTime": {
            "type": "string",
            "description": "Departure datetime.",
            "format": "date-time"
          },
          "arrivalTime": {
            "type": "string",
            "description": "Arrival datetime.",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "Scheduled",
              "Delayed",
              "Cancelled",
              "InAir",
              "Landed"
            ],
            "description": "Flight status (Scheduled, Delayed, Cancelled, InAir, Landed))"
          }
        },
        "required": [
          "flightNumber",
          "airline",
          "departureAirport",
          "arrivalAirport",
          "departureTime",
          "arrivalTime",
          "status"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "title": "ErrorResponse",
        "description": "Standard error response returned when a request cannot be processed.",
        "properties": {
          "error": {
            "type": "string",
            "description": "A short, machine-readable error code."
          },
          "message": {
            "type": "string",
            "description": "A human-readable description of the error."
          },
          "details": {
            "type": "array",
            "description": "Optional list of additional error details, such as validation messages.",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SearchKeys": {
        "type": "object",
        "title": "SearchKeys",
        "description": "Search criteria for filtering flights.",
        "properties": {
          "airline": {
            "type": "string",
            "description": "Search flights by airline."
          },
          "departureAirport": {
            "type": "string",
            "description": "Search flights by departure airport."
          },
          "arrivalAirport": {
            "type": "string",
            "description": "Search flights by arrival airport."
          },
          "fromDate": {
            "type": "string",
            "description": "Search flights by departure date from.",
            "format": "date"
          },
          "toDate": {
            "type": "string",
            "description": "Search flights by departure date to.",
            "format": "date"
          }
        },
        "x-internal": true
      }
    }
  }
}